mv cluster related funcs to misc_cluster
