+++ utils/__init__.py
@@ -3,6 +3,7 @@
 
 from .misc import *
 from .knn import *
+from .misc_cluster import *
 from .dataset import BasicDataset
 from .logger import create_logger
 from .search import faiss_search_knn
